Refactor folder structure to better match components
