+++ src/pages/MyMove/CreateOrEditMtoShipment.jsx
@@ -17,7 +17,7 @@ import {
 } from 'store/entities/selectors';
 import { AddressShape, SimpleAddressShape } from 'types/address';
 import { LocationShape } from 'types/index';
-import DateAndLocation from 'pages/MyMove/PPMBooking/DateAndLocation/DateAndLocation';
+import DateAndLocation from 'pages/MyMove/PPM/Booking/DateAndLocation/DateAndLocation';
 
 export class CreateOrEditMtoShipment extends Component {
   componentDidMount() {
